elf: fix dladdr call on newer rust
As reported in https://github.com/codyps/buildid/issues/11, newer rust
versions (at least 1.79) began failing to obtain the build-id on
elf-based platforms (linux, etc).

This occured because the value passed to `dladdr` became `1` instead of
a real address (TODO: confirm), possibly due to a change in how rust
treats `&fn()` types or how it treats chained pointer casts (`as *const
_ as *mut _`).

Avoid all this by using the `fn` directly to obtain the address and
eliminate the wildcard (`_`) pointer casts as they're not needed.